requirements.txt
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httpcore==0.16.3 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.23.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.9" and python_version < "4.0"
python-telegram-bot==20.0b0 ; python_version >= "3.9" and python_version < "4.0"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
